requirements.txt
fastapi[standard]
pydantic
requests
haystack-ai
qdrant-haystack
markdown-it-py mdit_plain
pypdf
sentence-transformers>=3.0.0
transformers
